moved my scratch code and scenes into dag dir
