+++ Assets/dag/BallController.cs
@@ -1,0 +1,29 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class BallController : MonoBehaviour {
+
+    public float speed = 1.0f;
+    Rigidbody rb;
+
+	// Use this for initialization
+	void Start () {
+        rb = gameObject.GetComponent<Rigidbody>();
+	}
+	
+	// Update is called once per frame
+	void Update () {
+		
+	}
+
+    private void FixedUpdate()
+    {
+        float moveHorizontal = Input.GetAxis("Horizontal");
+        float moveVertical = Input.GetAxis("Vertical");
+
+        Vector3 direction = new Vector3(moveHorizontal, 0.0f, moveVertical);
+
+        rb.AddForce(direction * speed);
+    }
+}
