+++ Assets/dag/class31may-tsts.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: ce37b385d97db40669fa89d74bb31cbf
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
